Update Divi.css
Craig updated lightbox to make background of the lightbox effect more transparent (was 90%, now 70%) so that it signals to the user that they're still on the same page. also made the close X icon lighter so it's more obvious, though you can close the image by clicking outside of it anyway.